unit tests working again
