+++ packages/testsweets/test/services/dynamic_keys_generator_service_test.dart
@@ -1,12 +1,12 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:testsweets/src/services/dynamic_keys_generator.dart';
 
-import '../helpers/test_helpers.dart';
+import '../helpers/dart_only_test_helpers.dart';
 
 void main() {
   group("DynamicKeysGeneratorService -", () {
-    setUp(registerServices);
-    tearDown(unregisterServices);
+    setUp(registerDartOnlyServices);
+    tearDown(unregisterDartOnlyServices);
 
     group("generateAutomationKeysForDynamicKey-", () {
       test(
