+++ packages/testsweets/test/services/widget_capture_service_test.dart
@@ -113,9 +113,6 @@ void main() {
         );
 
         final cloudFunctionsService = getAndRegisterCloudFunctionsService();
-        getAndRegisterTestSweetsConfigFileService(
-          valueFromConfigFileByKey: 'projectId',
-        );
 
         final service = WidgetCaptureService();
         await service.captureWidgetDescription(
